requirements.txt
alembic==1.11.2
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
dnspython==2.4.1
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.101.0
fastapi-cache2==0.2.1
fastapi-users==12.1.1
fastapi-users-db-sqlalchemy==6.0.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
passlib==1.7.4
pendulum==2.1.2
pycparser==2.21
pydantic==2.1.1
pydantic-core==2.4.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytzdata==2020.1
redis==4.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.23.2
zipp==3.16.2
